# backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "企业知识库问答系统 API",
    "version": "1.0.0",
...
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "本地开发"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "UserId": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-id",
        "description": "登录用户的 MongoDB _id"
      },
      "UserName": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-name",
        "description": "登录用户名"
      },
      "UserRole": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-role",
        "description": "用户角色：admin 或 user"
      }
    },
    "schemas": {
...
      "Category": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive"
            ]
          },
          "docCount": {
            "type": "integer"
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/admin/category-stats": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "知识库文档占比统计",
        "description": "按知识库分组统计文档数量，用于饼状图。当前无权限校验。",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "name": {
                            "type": "string"
                          },
                          "value": {
                            "type": "integer"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/api/admin/question-trend": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "提问趋势统计",
        "description": "返回最近 N 天每日提问数量，用于折线图。当前无权限校验。",
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "schema": {
              "type": "integer",
              "default": 7
            },
            "description": "统计天数（含今天）"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "date": {
                            "type": "string",
                            "example": "2026-07-16"
                          },
                          "count": {
                            "type": "integer"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/api/category": {
      "post": {
        "tags": [
          "Category"
        ],
        "summary": "新增知识库分类",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Category"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Category"
        ],
        "summary": "获取知识库分类列表",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "inactive"
              ]
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "in": "query",
            "name": "title",
            "schema": {
              "type": "string"
            },
            "description": "按名称搜索（title 或 name 均可）"
          },
          {
            "in": "query",
            "name": "name",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功，含分页和 docCount"
          }
        }
      }
    },
    "/api/category/{id}": {
      "get": {
        "tags": [
          "Category"
        ],
        "summary": "获取分类详情",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功"
          },
          "404": {
            "description": "知识库不存在"
          }
        }
      },
      "put": {
        "tags": [
          "Category"
        ],
        "summary": "更新知识库分类",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "inactive"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功"
          },
          "404": {
            "description": "知识库不存在"
          }
        }
      },
      "delete": {
        "tags": [
          "Category"
        ],
        "summary": "删除知识库分类",
        "description": "删除分类并将关联文档的 categoryId 设为 null",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          },
          "404": {
            "description": "知识库不存在"
          }
        }
      }
    },
    "/api/category/{id}/status": {
      "put": {
        "tags": [
          "Category"
        ],
        "summary": "切换分类状态",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "inactive"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "状态更新成功"
          },
          "400": {
            "description": "状态参数错误"
          },
          "404": {
            "description": "知识库不存在"
          }
        }
      }
    }
  }
}